+++ src/main.tsx
@@ -5,6 +5,7 @@ import { BrowserRouter } from "react-router-dom";
 import "./i18n/i18n";
 import App from "./App";
 import { LanguageProvider } from "./i18n/LanguageContext";
+import { TimezoneProvider } from "./time/TimezoneContext";
 import { ThemeProvider } from "./theme/ThemeContext";
 import GlobalStyle from "./theme/GlobalStyle";
 
@@ -15,10 +16,12 @@ ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
     <ThemeProvider>
       <LanguageProvider>
-        <GlobalStyle />
-        <BrowserRouter>
-          <App />
-        </BrowserRouter>
+        <TimezoneProvider>
+          <GlobalStyle />
+          <BrowserRouter>
+            <App />
+          </BrowserRouter>
+        </TimezoneProvider>
       </LanguageProvider>
     </ThemeProvider>
   </React.StrictMode>